Modfiy the linkTo module
