+++ src/libs/Navigation/dismissModal.ts
@@ -19,7 +19,7 @@ function dismissModal(
   }
 
   const state = navigationRef.getState();
-  const lastRoute = state.routes.at(-1);
+  const lastRoute = state.routes[state.routes.length - 1];
   switch (lastRoute?.name) {
     case NAVIGATORS.FULL_SCREEN_NAVIGATOR:
     case NAVIGATORS.LEFT_MODAL_NAVIGATOR:
