+++ src/libs/Navigation/linkTo/types.ts
@@ -1,0 +1,11 @@
+type ActionPayloadParams = {
+  screen?: string;
+  params?: unknown;
+  path?: string;
+};
+
+type ActionPayload = {
+  params?: ActionPayloadParams;
+};
+
+export type {ActionPayload, ActionPayloadParams};
